feat(ami): pure terraform impl of nix ami lookup
This is a pure terraform implementation of the nix AMI lookup. It has
been validated to find the same AMIs for 14.12+ as the nix AMI list.

The single issue is that latest might take the very latest AMI instead
of one marked in the AMI list. Unless NixOS starts tagging AMIs in a
way that would allow differentiation between rc/test builds there is the
chance that latest can break. However, not pinning versions is generally
bad practice so this may be a minor concern.

NOTE: This errors when trying to find AMIs for 14.04 because it appears
those AMIs no longer exist.